Move Singleton & AbstractFactory to new folders
