+++ cpp17Play/design_pattern/CMakeLists.txt
@@ -1,16 +1,3 @@
+add_subdirectory(creational_patterns)
 
-add_library(
-    AbstractFactory
-    INTERFACE
-    "AbstractFactory.hpp"
-)
-
-add_library(
-    Singleton
-    INTERFACE
-    "Singleton.hpp"
-    # SHARED
-    # "Singleton.cpp"
-)
-
-add_subdirectory(test)
+add_subdirectory(structural_patterns)
