+++ cpp17Play/design_pattern/creational_patterns/CMakeLists.txt
@@ -1,0 +1,16 @@
+
+add_library(
+    AbstractFactory
+    INTERFACE
+    "AbstractFactory.hpp"
+)
+
+add_library(
+    Singleton
+    INTERFACE
+    "Singleton.hpp"
+    # SHARED
+    # "Singleton.cpp"
+)
+
+add_subdirectory(test)
